% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{accuracy}
\alias{accuracy}
\title{Classification Accuracy}
\usage{
accuracy(tbl)
}
\arguments{
\item{tbl}{table with predicted and reference classes correctly matched}
}
\value{
classification accuracy, given as a proportion
}
\description{
Given a sorted table, the accuracy is calculated based on proportion
correctly classified.
}
\details{
The table must have rows and columns corresponding to the same classes. Then
the classification accuracy is computed as the sum of the diagonal entries
divided by the sum of all entries in the confusion matrix comparing the
predicted and reference classes.
}
\examples{
set.seed(1)
x <- matrix(rbinom(16, 20, 0.4), nrow = 4)
accuracy(x)
}
\author{
Derek Chiu
}

