% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakE.R
\name{breakE}
\alias{breakE}
\title{Title function to create a cluster matrix for all algorithms in ConClust during 1 of the repetitions}
\usage{
breakE(E, methods, col_ind)
}
\arguments{
\item{E}{: ensemble out of crEnsemble using scheme=3}

\item{methods}{: algorithm names as those in ConClust}

\item{col_ind}{: indicator of which repetition. Range is 1 to dim(E)[2]}
}
\value{
a 2D matrix
}
\description{
Title function to create a cluster matrix for all algorithms in ConClust during 1 of the repetitions
}

