% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPosInt.R
\name{checkPosInt}
\alias{checkPosInt}
\title{Title function to check if a number is a positive integer}
\usage{
checkPosInt(x)
}
\arguments{
\item{x}{: a number}
}
\value{
TRUE if x is a positive integer and FALSE otherwise
}
\description{
Title function to check if a number is a positive integer
}

