% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVec.R
\name{checkVec}
\alias{checkVec}
\title{Title convert one cluster result to another cluster with same algorithm but different labeling}
\usage{
checkVec(numClust, v2, vec)
}
\arguments{
\item{numClust}{: number of clusters}

\item{v2}{: reference vector}

\item{vec}{: vector to be converted}
}
\value{
converted vec
}
\description{
Title convert one cluster result to another cluster with same algorithm but different labeling
}
\examples{
checkVec(4,c(4,2,1,3),c(1,3,2,4)) returns 3,4,2,1
}

