% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleval.R
\name{cleval}
\alias{cleval}
\title{Title : function to compute validity scores for clustering results}
\usage{
cleval(X, CR, methods, truelabels)
}
\arguments{
\item{X}{: a dataset, rows correspond to observations, columns correspond to variables (excluding class labels!)}

\item{CR}{: a matrix of clustering resutls (excluding row and column headers !)}

\item{methods}{: a vector of legend strings for bar plots}

\item{truelabels}{: vector for true labels for each data point}
}
\value{
V: matrix of cluster validity scores
}
\description{
Title : function to compute validity scores for clustering results
}

