% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_clust.R
\name{clHC}
\alias{clHC}
\title{Title: function to perform final clustering using hierarchical algorithms (single linkage-SL, complete linkage-CL, average linkage-AL)}
\usage{
clHC(S, K)
}
\arguments{
\item{S}{: N by N similarity matrix}

\item{K}{: preferred number of clusters}
}
\value{
CR: an N by 3 matrix of clustering results from SL, CL, and AL
}
\description{
Title: function to perform final clustering using hierarchical algorithms (single linkage-SL, complete linkage-CL, average linkage-AL)
}

