% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Title function to create clusters using distance criterion from a hierarchical cluster tree}
\usage{
cluster(Z, maxclust)
}
\arguments{
\item{Z}{: matrix representing a hierarchical cluster tree}

\item{maxclust}{: desired number of clusters}
}
\value{
resultT: clustering result
}
\description{
Title function to create clusters using distance criterion from a hierarchical cluster tree
}

