% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDuck-package.r
\docType{package}
\name{ClusterDuck-package}
\alias{ClusterDuck-package}
\title{ClusterDuck}
\description{
The ClusterDuck package performs Consensus Clustering, an ensemble clustering
algorithm that pools together results from different clustering realizations.
Each realization uses a unique combination of observations and clustering
algorithms.
}

