% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusternum.R
\name{clusternum}
\alias{clusternum}
\title{Title function to assign leaves under cluster c to c, equivalent to MATLAB function clusternum}
\usage{
clusternum(X, resultT, k, c)
}
\arguments{
\item{c}{}
}
\description{
Title function to assign leaves under cluster c to c, equivalent to MATLAB function clusternum
}

