% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colMax}
\alias{colMax}
\alias{colMin}
\title{Form Row and Column Extremes}
\usage{
colMax(x, na.rm = TRUE)

colMin(x, na.rm = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{na.rm}{logical; Should missing values be omitted from consideration?}
}
\value{
a vector containing the maximums or minimums for each column in
  \code{x}
}
\description{
Form row and column maxs and mins for matrices
}
\examples{
a <- matrix(1:9, ncol = 3)
colMax(a)
colMin(a)
}

