% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConClust-package.r
\docType{package}
\name{ConClust-package}
\alias{ConClust-package}
\title{ConClust}
\description{
The ConClust package performs Consensus Clustering, an ensemble clustering
algorithm that pools together results from different clustering realizations.
Each realization uses a unique combination of observations and clustering
algorithms.
}

