% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conClustEnsemble.R
\name{conClustEnsemble}
\alias{conClustEnsemble}
\title{Title : function to generate a cluster ensemble from the data set X using K-means algorithm}
\usage{
conClustEnsemble(X, k, reps = 1000)
}
\arguments{
\item{X}{: a data set, rows are observations, columns are variables}

\item{k}{: number of clusters in the base clusterings}

\item{M}{: number of base clusterings in the ensemble}

\item{scheme}{: cluster ensemble generating scheme (1=fixed k, 2=random k)}
}
\value{
E: matrix of cluster ensemble
}
\description{
Title : function to generate a cluster ensemble from the data set X using K-means algorithm
}

