% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_class.R
\name{consensus_class}
\alias{consensus_class}
\title{Consensus class using hierarchical clustering}
\usage{
consensus_class(x, k, method = "average", names = NULL)
}
\arguments{
\item{x}{a \code{\link{consensus_matrix}}}

\item{k}{number of clusters}

\item{method}{linkage type for hierarchical clustering. Defaults to
"average". See \code{\link[stats]{hclust}} for details.}

\item{names}{vector of sample labels. Default is \code{NULL}, so that
the names are obtained from the \code{hclust} dendrogram labels.}
}
\value{
cluster assignments for the consensus class
}
\description{
Performs hierarchical clustering on a consensus matrix to obtain consensus
class labels.
}
\examples{
set.seed(2)
x <- replicate(100, rbinom(100, 4, 0.2))
cm <- consensus_matrix(x)
consensus_class(cm, k = 3)
}
\author{
Derek Chiu
}
\seealso{
Other consensus functions: \code{\link{consensus_combine}},
  \code{\link{consensus_confmat}},
  \code{\link{consensus_matrix}},
  \code{\link{consensus_summary}}
}

