% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cluster.R
\name{consensus_cluster}
\alias{consensus_cluster}
\title{Consensus clustering}
\usage{
consensus_cluster(data, nk = 2:4, pItem = 0.8, reps = 1000,
  algorithms = NULL, nmf.method = c("brunet", "lee"),
  distance = "euclidean", prep.data = c("full", "sampled", "none"),
  min.sd = 1, progress = TRUE, seed.nmf = 123456, seed.data = 1,
  save = FALSE, file.name = "CCOutput", time.saved = FALSE)
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}

\item{pItem}{proportion of items to be used in subsampling within an 
algorithm}

\item{reps}{number of subsamples}

\item{algorithms}{vector of clustering algorithms for performing consensus 
clustering. Must be any number of the following: "nmf", 
"hc", "diana", "km", "pam", "ap", "sc", "gmm", "block". See details. Can
use a custom clustering algorithm. See example.}

\item{nmf.method}{specify NMF-based algorithms to run. By default the 
"brunet" and "lee" algorithms are called. See \code{\link[NMF]{nmf}} for
details.}

\item{distance}{a vector of distance functions. Defaults to "euclidean". Can 
use a custom distance function. See example.}

\item{prep.data}{Prepare the data on the "full" dataset (default), the
"sampled" dataset, or "none".}

\item{min.sd}{minimum standard deviation threshold for \code{prep.data}}

\item{progress}{logical; should a progress bar be displayed?}

\item{seed.nmf}{random seed to use for NMF-based algorithms}

\item{seed.data}{seed to use to ensure each algorithm operates on the same set
of subsamples}

\item{save}{logical; if \code{TRUE}, the returned object will be saved at 
each iteration as well as at the end.}

\item{file.name}{file name of the written object}

\item{time.saved}{logical; if \code{TRUE}, the date saved is appended to the 
file name. Only applicable when \code{dir} is not \code{NULL}.}
}
\value{
An array of dimension \code{nrow(x)} by \code{reps} by 
  \code{length(algorithms)} Each slice of the array is a matrix showing 
  consensus clustering results for algorithms. The matrices have a row for
  each sample, and a column for each subsample. Each entry represents a class
  membership.
}
\description{
Runs consensus clustering across subsamples of the data, clustering 
algorithms, and cluster sizes.
}
\details{
The clustering algorithms provided are:
\itemize{
  \item{"nmf": }{Nonnegative Matrix Factorization using Kullback-Leibler Divergence} 
  \item{"nmfEucl": }{Nonnegative Matrix Factorization using Euclidean distance}
  \item{"hc": }{Hierarchical Clustering}
  \item{"diana": }{DIvisive ANAlysis Clustering}
  \item{"km": }{K-Means Clustering}
  \item{"pam": }{Partition Around Mediods}
  \item{"ap": }{Affinity Propagation}
  \item{"sc": }{Spectral Clustering using Radial-Basis kernel function}
  \item{"gmm": }{Gaussian Mixture Model using Bayesian Information Criterion on EM algorithm} 
  \item{"block": }{Biclustering using a latent block model}
}

The \code{nmf.method} defaults are "brunet" (Kullback-Leibler divergence) and
"lee" (Euclidean distance).

The progress bar increments for every unit of \code{reps}.
}
\examples{
data(hgsc)
dat <- t(hgsc[, -1])

# Custom distance function
manh <- function(x) {
  stats::dist(x, method = "manhattan")
}

# Custom clustering algorithm
agnes <- function(d, k) {
  return(as.integer(stats::cutree(cluster::agnes(d, diss = TRUE), k)))
}

cc <- consensus_cluster(dat, reps = 5, algorithms = c("pam", "agnes"),
distance = c("euclidean", "manh"))
str(cc)
}
\author{
Derek Chiu, Aline Talhouk
}
