% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{consensus_cluster_old}
\alias{consensus_cluster_old}
\title{Consensus clustering over samples and algorithms}
\usage{
consensus_cluster_old(data, nk = 2:4, pItem = 0.8, reps = 1000,
  algorithms = NULL, parallel = NULL, ncores = NULL, progress = TRUE,
  seed = 123456, seed.alg = 1, min.sd = 1, save = FALSE,
  file.name = "CCOutput", time.saved = FALSE)
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}

\item{pItem}{proportion of items to be used in subsampling within an 
algorithm}

\item{reps}{number of subsamples}

\item{algorithms}{vector of clustering algorithms for performing consensus 
clustering. Must be any number of the following: "nmfDiv", "nmfEucl", 
"hcAEucl", "hcDianaEucl", "kmEucl", "kmSpear", "pamEucl", "pamSpear", 
"apEucl", "scRbf", "gmmBIC", "biclust". See details.}

\item{parallel}{logical; if \code{TRUE}, the function registers a parallel 
backend from the \code{doParallel} package and runs a foreach loop. By 
default, \code{parallel = NULL} means the function determines, based on 
other parameters, whether to use parallel or not (e.g. number of
\code{reps}, availability of multiple cores).}

\item{ncores}{number of CPU cores to use for parallel computation. Default 
uses all available cores.}

\item{progress}{logical; should a progress bar be displayed?}

\item{seed}{random seed to use for NMF-based algorithms}

\item{seed.alg}{seed to use to ensure each algorithm operates on the same set
of subsamples}

\item{min.sd}{minimum standard deviation threshold. See details.}

\item{save}{logical; if \code{TRUE}, the returned object will be saved at 
each iteration as well as at the end.}

\item{file.name}{file name of the written object}

\item{time.saved}{logical; if \code{TRUE}, the date saved is appended to the 
file name. Only applicable when \code{dir} is not \code{NULL}.}
}
\value{
An array of dimension \code{nrow(x)} by \code{reps} by 
  \code{length(algorithms)} Each slice of the array is a matrix showing 
  consensus clustering results for algorithms. The matrices have a 
  row for each sample, and a column for each subsample. Each entry represents
  a class membership.
}
\description{
Generates multiple runs for consensus clustering among replicated subsamples 
of a dataset as well as across different clustering algorithms.
}
\details{
The clustering algorithms provided are:
\itemize{ 
\item{"nmfDiv": }{Nonnegative Matrix Factorization using Kullback-Leibler Divergence}
\item{"nmfEucl": }{Nonnegative Matrix Factorization using Euclidean distance}
\item{"hcAEucl": }{Hierarchical Clustering using Average linkage and Euclidean distance}
\item{"hcDianaEucl": }{Divisive Hierarchical Clustering using Euclidean distance}
\item{"kmEucl": }{K-Means using Euclidean distance}
\item{"kmSpear": }{K-Means using Spearman distance} 
\item{"pamEucl": }{Partition Around Mediods using Euclidean distance} 
\item{"pamSpear": }{Partition Around Mediods using Spearman distance} 
\item{"apEucl": }{Affinity Propagation using Euclidean distance} 
\item{"scRbf": }{Spectral Clustering using Radial-Basis kernel function} 
\item{"gmmBIC": }{Gaussian Mixture Model using Bayesian Information Criterion on EM algorithm}
\item{"biclust": }{Biclustering using a latent block model} 
}

The \code{min.sd} argument is used to filter the feature space for only 
highly variable features. Only features with a standard deviation across all 
samples greater than \code{min.sd} will be used.

The progress bar for \code{parallel = FALSE} increments for every unit of 
\code{reps}, where as for \code{parallel = TRUE}, the increments are for 
\code{algorithms}. Additionally, when the progress bar is used in 
conjunction with parallel computation, each worker outputs the attached 
packages (and potential masking of functions).
}
\author{
Derek Chiu, Aline Talhouk
}

