% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_confmat.R
\name{consensus_confmat}
\alias{consensus_confmat}
\title{Confusion matrix between consensus clusters and true classes}
\usage{
consensus_confmat(cl.cons, cl.true, pred.lab = "Prediction",
  ref.lab = "Reference")
}
\arguments{
\item{cl.cons}{vector of cluster assignments}

\item{cl.true}{true class labels}

\item{pred.lab}{label for the predicted clusters}

\item{ref.lab}{label for the reference classes}
}
\value{
A confusion matrix with the predicted cluster assignments compared to
  the reference true class labels.
}
\description{
Compares the consensus class to the true class partitions.
}
\examples{
set.seed(2)
a <- sample(1:4, 100, replace = TRUE)
b <- sample(1:4, 100, replace = TRUE)
consensus_confmat(a, b)
}
\author{
Derek Chiu
}
\seealso{
Other consensus.functions: \code{\link{consensus_class}},
  \code{\link{consensus_combine}},
  \code{\link{consensus_matrix}},
  \code{\link{consensus_summary}}
}

