% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contingency.R
\name{Contingency}
\alias{Contingency}
\title{Title Part of valid_RandIndex}
\usage{
Contingency(Mem1, Mem2)
}
\arguments{
\item{Mem1}{: a vector}

\item{Mem2}{: a vector}
}
\value{
See matlab valid_RandIndex.m in LinkCluE
}
\description{
Title Part of valid_RandIndex
}

