% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{cts}
\alias{cts}
\title{Title : fucntion for computing connected triple based similarity matrix}
\usage{
cts(E, dc)
}
\arguments{
\item{E}{: matrix of cluster ensemble}

\item{dc}{: decay factor, [0, 1]}
}
\value{
S: CTS matrix
}
\description{
Title : fucntion for computing connected triple based similarity matrix
}

