% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Diverse Clustering Ensemble}
\usage{
dice(data, nk, R = 10, algorithms = c("hcAEucl", "kmEucl", "scRbf",
  "gmmBIC"), consensusFUNS = c("kmodes", "CSPA", "majority", "LCE"),
  sim.mat = c("cts", "srs", "asrs"), trim = FALSE, reweigh = FALSE,
  evaluate = TRUE, refClass = NULL, progress = TRUE)
}
\arguments{
\item{data}{matrix with rows as observations, columns as variables}

\item{nk}{a range of cluster sizes (or a single value)}

\item{R}{number of data subsamples to generate. See \code{\link{ConClust}} 
for details.}

\item{algorithms}{clustering algorithms to be used in the ensemble. Current 
options are "nmfDiv", "nmfEucl", "hcAEucl", "hcDianaEucl", "kmEucl", 
"kmSpear", "pamEucl", "pamSpear", "apEucl", "scRbf", "gmmBIC", and 
"biclust". See \code{\link{ConClust}} for details.}

\item{consensusFUNS}{consensus functions to use. Current options are "kmodes"
(k-modes), "majority" (majority voting), "CSPA" (hierarchical clustering), 
"LCE" (linkage clustering ensemble)}

\item{sim.mat}{type of similarity matrix. One of "cts", "srs", "asrs. See
\code{\link{LCE}} for details.}

\item{trim}{logical; if \code{TRUE}, the number of algorithms in 
\code{algorithms} is reduced based on internal validity index performance 
prior to consensus clustering by \code{consensusFUNS}. Defaults to 
\code{FALSE}.}

\item{reweigh}{logical; if \code{TRUE}, algorithms are reweighted based on 
internal validity index performance after trimming. Well-performing 
algorithms are given higher weight prior to consensus clustering by 
\code{consensusFUNS}. Defaults to \code{FALSE}. Ignored if \code{trim = 
FALSE}.}

\item{evaluate}{logical; if \code{TRUE} (default), validity indices are 
returned. Internal validity indices are always computed. If \code{refClass}
is not \code{NULL}, then external validity indices will also be computed.}

\item{refClass}{reference class; a vector of length equal to the number of 
observations.}

\item{progress}{logical; if \code{TRUE} (default), progress bar is shown.}
}
\value{
A final clustering assignment from the diverse clustering ensemble
  method.
}
\description{
Runs consensus clustering across subsamples, algorithms, and number of 
clusters (k).
}
\examples{
library(dplyr)
data(hgsc)
dat <- t(hgsc[, -1])
refClass <- data.frame(initCol = rownames(dat)) \%>\%
tidyr::separate(initCol,
                into = c("patientID", "Class"),
                sep = "_") \%>\% 
  magrittr::use_series(Class) \%>\% 
  factor() \%>\% 
  as.integer()
dice.obj <- dice(dat, nk = 4, algorithms = c("hcAEucl", "hcDianaEucl",
"pamEucl", "pamSpear"), consensusFUNS = c("kmodes", "majority", "LCE"), trim
= TRUE, refClass = refClass)
str(dice.obj)
}
\author{
Aline Talhouk, Derek Chiu
}

