% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diceR-package.R
\docType{package}
\name{diceR-package}
\alias{diceR}
\alias{diceR-package}
\title{diceR: Diverse Cluster Ensemble in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Performs cluster analysis using an ensemble clustering framework, Chiu & Talhouk (2018) \doi{10.1186/s12859-017-1996-y}. Results from a diverse set of algorithms are pooled together using methods such as majority voting, K-Modes, LinkCluE, and CSPA. There are options to compare cluster assignments across algorithms using internal and external indices, visualizations such as heatmaps, and significance testing for the existence of clusters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlineTalhouk/diceR/}
  \item \url{https://alinetalhouk.github.io/diceR/}
  \item Report bugs at \url{https://github.com/AlineTalhouk/diceR/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Derek Chiu \email{dchiu@bccrc.ca}
  \item Aline Talhouk \email{a.talhouk@ubc.ca}
}

Other contributors:
\itemize{
  \item Johnson Liu \email{gliu@bccrc.ca} [contributor, compiler]
}

}
\keyword{internal}
