% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{E_LCE}
\alias{E_LCE}
\title{Example data for Linkage Cluster Ensemble (LCE)}
\format{A matrix with 100 rows (samples) and 10 columns (cluster assignments).}
\usage{
E_LCE
}
\description{
The data set is intended to be used with the LCE family of functions
}
\keyword{datasets}

