% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalConCl.R
\name{evalConCl}
\alias{evalConCl}
\title{Title function to evaluate ensemble result using ConClust}
\usage{
evalConCl(X, E, R, truelabels = NULL, K, algNames = c("nmfDiv", "nmfEucl",
  "hcAEucl", "hcDianaEucl", "kmEucl", "kmSpear", "pamEucl", "pamSpear",
  "apEucl", "scRbf", "gmmBIC", "biclust"), methods = c("CTS-SL", "CTS-CL",
  "CTS-AL", "SRS-SL", "SRS-CL", "SRS-AL", "ASRS-SL", "ASRS-CL", "ASRS-AL"),
  mNames = c("cts", "srs", "asrs"), dcs = c(0.8, 0.8, 0.8))
}
\arguments{
\item{X}{: original data, rows are samples and columns are genes}

\item{E}{: output from crEnsemble in LinkCluE using scheme=3 (ConClust)}

\item{R}{: number of iterations for simrank algorithm (don't need for CTS)}

\item{truelabels}{:optional, known cluster labels for each data points}

\item{K}{:number of clusters in the consensus functions}

\item{algNames}{: names of algorithms used in ConClust}

\item{methods}{: default: "CTS-SL","CTS-CL","CTS-AL","SRS-SL","SRS-CL","SRS-AL","ASRS-SL","ASRS-CL","ASRS-AL"}

\item{mNames}{: methods for creating similarity matrices}

\item{dcs}{: decay constants vector}
}
\description{
Title function to evaluate ensemble result using ConClust
}

