% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_validity.R, R/internal_validity.R
\name{external_validity}
\alias{ev_accuracy}
\alias{ev_rand}
\alias{external_validity}
\alias{iv_db_dunn}
\alias{iv_sumsq}
\title{External validity indices}
\usage{
ev_accuracy(pred.lab, ref.lab)

ev_rand(pred.lab, ref.lab)

iv_db_dunn(x, labels)

iv_sumsq(data, labels, k)
}
\arguments{
\item{pred.lab}{labels generated by classifier}

\item{ref.lab}{true labels for the observations}

\item{k}{number of clusters}
}
\value{
\code{ev_accuracy} returns the classification accuracy

\code{ev_rand} returns a list with elements
  \item{AR}{adjusted Rand index}
  \item{RI}{unadjusted Rand index}
  \item{MI}{Mirkin's index}
  \item{HI}{Hubert's index}

\code{iv_db_dunn} returns a list with elements
  \item{DB}{Davies-Bouldin score}
  \item{Dunn}{Dunn score}

\code{iv_sumsq} returns a list with elements 
  \item{W}{within group sum of squares and cross-products}
  \item{B}{between group sum of squares and cross-products}
  \item{T}{total sum of squares and cross-products}
  \item{Sintra}{centroid diameter}
  \item{Sinter}{linkage distance}
}
\description{
\strong{E}xternal \strong{v}alidity indices compare a predicted clustering 
result with a reference class or gold standard.

\code{iv_db_dunn} computes the Davies-Bouldin index and Dunn index
}
\details{
\code{ev_accuracy} computes the classification accuracy for a clustering
result

\code{ev_rand} calculates Rand indices (and their variants)to
  compare two partitions

\code{iv_sumsq} computes the within group, between group, and total
  sum of squares and cross-products
}
\examples{
data("E")
ev_accuracy(c(4, 4, 4, 4), c(4, 2, 4, 4))
ev_accuracy(c(4, 4, 4, 4, 8), c(4, 2, 4, 4, 6))
ev_rand(E[, 1], E[, 2])
}

