% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FGT}
\alias{FGT}
\title{Example dataset FGT}
\format{A data frame with 100 rows and 1 column}
\usage{
FGT
}
\description{
The single column has 25 elements evenly placed in 4 groups.
}
\keyword{datasets}

