% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_c2_valid_RandIndex}
\alias{final_c2_valid_RandIndex}
\title{Example dataset final_c2_valid_RandIndex}
\format{A vector of length 200}
\usage{
final_c2_valid_RandIndex
}
\description{
This vector is intended to be tested in \code{ev_rand}
}
\keyword{datasets}

