% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fixedKmeansTCGA_CR}
\alias{fixedKmeansTCGA_CR}
\title{TCGA example data}
\format{A matrix with 489 rows and 9 columns}
\usage{
fixedKmeansTCGA_CR
}
\description{
The columns are intended to represent different cluster assignments.
}
\keyword{datasets}

