% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fixedKmeansTCGA_V}
\alias{fixedKmeansTCGA_V}
\title{TCGA example data}
\format{A matrix with 3 rows and 9 columns}
\usage{
fixedKmeansTCGA_V
}
\description{
The data is assessed using different similarity measures and external
validity measures.
}
\keyword{datasets}

