% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnMin.R
\name{getColumnMin}
\alias{getColumnMin}
\title{Title: function to find the minimum in each column of a mtrix}
\usage{
getColumnMin(M)
}
\arguments{
\item{M}{: input matrix}
}
\value{
vec: vector containing the minimum in each column of input matrix
}
\description{
Title: function to find the minimum in each column of a mtrix
}

