% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRowColNumbers.R
\name{getRowColNumbers}
\alias{getRowColNumbers}
\title{Title: function to find the locations of an enetry in a matrix}
\usage{
getRowColNumbers(M, toFind)
}
\arguments{
\item{M}{: input matrix}

\item{toFind}{: number to be found}
}
\value{
: rows: vector containing the row numbers of toFind in M
         cols: vector containing the column numbers of toFind in M
}
\description{
Title: function to find the locations of an enetry in a matrix
}

