% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRowOfMin.R
\name{getRowOfMin}
\alias{getRowOfMin}
\title{Title function to get row number of the minimum in each column of a matrix}
\usage{
getRowOfMin(M)
}
\arguments{
\item{M}{: a matrix}
}
\value{
vector containing row numbers of the minimum in each column of a matrix
}
\description{
Title function to get row number of the minimum in each column of a matrix
}

