% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_cdf}
\alias{graph_cdf}
\title{Cumulative distribution function graphs}
\usage{
graph_cdf(x)
}
\arguments{
\item{x}{an object from \code{consensus_combine} using option \code{element
= "matrix"}. A list of consensus matrices from different clustering
algorithms is the input.}
}
\value{
a ggplot showing the cdfs for different algortihms
}
\description{
Graph CDF for consensus matrices from different algorithms
}
\examples{
# Consensus clustering for 3 algorithms
set.seed(911)
x <- matrix(rnorm(1000), nrow = 10)
CC1 <- ConClust(x, k = 4, reps = 10,
method = c("hcAEucl", "apEucl", "gmmBIC"), save = FALSE)
CC1.summ <- consensus_summary(CC1, k = 4)
y <- consensus_combine(CC1.summ, element = "matrix")

# Plot CDF
p <- graph_cdf(y)
p

# Change y label and add colours
p + labs(y = "Probability") + stat_ecdf(aes(colour = Method)) +
scale_color_brewer(palette = "Set2")
}
\author{
Derek Chiu
}

