% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_delta_area}
\alias{graph_delta_area}
\title{Delta Area graph}
\usage{
graph_delta_area(x)
}
\arguments{
\item{x}{an object from \code{\link{ConClust}}}
}
\description{
Calculate the relative change in area under CDF curve between algorithms
}

