% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_heatmap}
\alias{graph_heatmap}
\title{Consensus matrix heatmaps}
\usage{
graph_heatmap(x, main = NULL, ...)
}
\arguments{
\item{x}{an object from \code{\link{ConClust}}}

\item{main}{heatmap title}

\item{...}{additional arguments to \code{\link[gplots]{heatmap.2}}}
}
\value{
consensus matrix heatmaps for each algorithm in \code{x}
}
\description{
Consensus matrix heatmaps
}
\author{
Derek Chiu
}

