% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgsc.R
\docType{data}
\name{hgsc}
\alias{hgsc}
\title{High grade serous cancer data from TCGA used to classify subtypes}
\format{A data frame with 321 rows and 490 variables}
\usage{
hgsc
}
\description{
There are 321 genes and 489 samples. The first column is a gene identifier.
}
\keyword{datasets}

