% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{hooks}
\alias{agnes_hook}
\alias{diana_hook}
\alias{hooks}
\title{Hook Functions}
\usage{
agnes_hook(d, k)

diana_hook(d, k)
}
\arguments{
\item{d}{distance matrix}

\item{k}{scalar indicating number of clusters to cut tree into}
}
\value{
clustering assignment from \code{agnes} or \code{diana}
}
\description{
Custom hook functions for AGglomerative NESting and DIvisive ANAlysis
clustering algorithms
}
\details{
The hierarchical structure returned using \code{agnes} is equivalent to using
\code{\link{hclust}}.
}
\examples{
library(cluster)
data(votes.repub)
d <- dist(votes.repub)
agnes_hook(d, k = 2)
diana_hook(d, k = 2)
}
\author{
Derek Chiu
}

