% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validity.R
\name{internal_validity}
\alias{internal_validity}
\alias{iv_compactness}
\title{Internal validity indices}
\usage{
iv_compactness(x, labels)
}
\arguments{
\item{x}{a dataset with rows as observations, columns sariables}

\item{labels}{a vector of cluster labels from a clustering result}
}
\value{
\code{iv_compactness} returns the compactness score
}
\description{
\strong{I}nternal \strong{v}alidity indices are agnostic to any reference
clustering results. They calculate cluster performance on the basis of 
compactness and separability.
}
\details{
\code{iv_compactness} computes the compactness validity index for a
clustering result
}
\examples{
data("FGD", "E", "FGT", "FGD")
iv_compactness(FGD, E[, 1])
iv_db_dunn(FGD, FGT)
iv_sumsq(data = FGD, labels = FGT, k = 4)
}

