% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{knn_impute}
\alias{knn_impute}
\title{K-Nearest Neighbours Imputation}
\usage{
knn_impute(res, data, seed = 123456)
}
\arguments{
\item{res}{a vector representing a column of the output from \code{ConClust}}

\item{data}{data matrix with samples as rows and genes as columns}

\item{seed}{random seed for reproducibility}
}
\value{
results matrix with missing values filled in
}
\description{
Use K-Nearest Neighbours to fill in samples not clustered due to bootstrap
resampling in consensus clustering.
}
\details{
The number of neighbours considered is 5 (default) and the minimum vote for
definite decision is 3 (default).
}
\examples{
# Generate data from ConClust
set.seed(2)
x <- replicate(100, rnorm(100))
CC <- ConClust(x, nc = 4, reps = 10, method = "scRbf")
orig.cl <- CC[, 1, 1, 1]
sum(is.na(orig.cl))

# Still missing values because there is doubt in the voting for some cases
impute.cl <- knn_impute(orig.cl, x)
sum(is.na(impute.cl))
}
\author{
Aline Talhouk
}

