% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_clust.R
\name{link_clust}
\alias{link_clust}
\title{Title function to take many clustering results and create a cluster ensemble}
\usage{
link_clust(E, dcCTS = 0.8, dcSRS = 0.8, dcASRS = 0.8, R = 10,
  is.relabelled = TRUE)
}
\arguments{
\item{E}{: matrix of clustering resutls}

\item{dcCTS}{: decay constant for CTS matrix}

\item{dcSRS}{: decay constant for SRS matrix}

\item{dcASRS}{: decay constant for ASRS matrix}

\item{R}{: number of repetitions for SRS matrix}

\item{is.relabelled}{: boolean representing whether input E is relabelled}
}
\value{
a list containing final clustering result using k modes, majority vote, and hierarchical clustering(average, single, and complete linkages)
}
\description{
Title function to take many clustering results and create a cluster ensemble
}

