% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage}
\alias{linkage}
\title{Create hierarchical cluster tree}
\usage{
linkage(Y, method)
}
\arguments{
\item{Y}{a distance vector}

\item{method}{linkage method; can be "average", "single", or "complete"}
}
\value{
matrix denoting hierarchical cluster tree
}
\description{
Create hierarchical cluster tree
}
\examples{
set.seed(1)
E<-matrix(rep(sample(1:4,1000,replace = TRUE)),nrow=100,byrow=FALSE)
lk<-linkage(stod(cts(E,0.8)),"average")
}
\author{
Johnson Liu
}
\references{
function linkage in MATLAB for average, single, and complete linkage
}

