% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkclueMain.R
\name{linkclueMain}
\alias{linkclueMain}
\title{Title function to perform link-based cluster ensemble for a dataset}
\usage{
linkclueMain(X, M = NULL, k, scheme, K, dcCTS = 0.8, dcSRS = 0.8, R = 5,
  dcASRS = 0.8, truelabels = NULL, reps = 1000)
}
\arguments{
\item{X}{: a dataset whose rows are observations and columns are variables}

\item{M}{: preferred number of base clusterings in the ensemble}

\item{k}{: preferred number of clusters in the base clusterings}

\item{scheme}{: cluster gensemble generating scheme (1=fixed, 2=random k)}

\item{K}{: number of clusters in the consensus functions}

\item{dcCTS}{: decay factor for CTS method}

\item{dcSRS}{; decay factor for SRS method}

\item{R}{: number of iterations for simrank algorithm (don't need for CTS)}

\item{dcASRS}{: decay factor for ASRS method}

\item{truelabels}{: optional, known cluster labels for each data points}

\item{reps:}{number of repetitions if scheme is 3 for conclust clustering}
}
\description{
Title function to perform link-based cluster ensemble for a dataset
}

