% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LT}
\alias{LT}
\title{Example dataset LT}
\format{A data frame with 72 rows and 1 column}
\usage{
LT
}
\description{
The single column has 24 elements in group 1 and 48 elements in group 2.
}
\keyword{datasets}

