% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majVote.R
\name{majVote}
\alias{majVote}
\title{Title function to perform major vote on a 3D array with NA entries}
\usage{
majVote(arr)
}
\arguments{
\item{arr}{: input character array with NAs}
}
\value{
a 3D numeric array with NAs in original array voted
}
\description{
Title function to perform major vote on a 3D array with NA entries
}

