% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigclust.R
\name{my_sigclust}
\alias{my_sigclust}
\title{Significant Testing of Clustering Results}
\usage{
my_sigclust(x, k, nsim, nrep = 1, labflag = 0, label = 0, icovest = 2)
}
\arguments{
\item{x}{data matrix, samples are rows and features are columns}

\item{k}{cluster size to test against}

\item{nsim}{number of simulations}

\item{nrep}{number of replicates}

\item{labflag}{See \code{\link[sigclust]{sigclust}} for details.}

\item{label}{true class label}

\item{icovest}{type of covariance matrix estimation}
}
\value{
An object of class \code{sigclust}. See
  \code{\link[sigclust]{sigclust}} for details.
}
\description{
Uses the SigClust K-Means algorithm to assess significance of clustering
results.
}
\details{
This function is a wrapper around the original
\code{\link[sigclust]{sigclust}}, except that an additional param \code{k} is
added for the case when k > 2.
}
\examples{
data(hgsc)
dat <- t(hgsc[, -1])[1:100, 1:50]
set.seed(1)
nk <- 4
label <- sample(seq_len(nk), nrow(dat), replace = TRUE)
cc <- consensus_cluster(dat, nk = 4, reps = 5, algorithms = "pam", progress =
FALSE)
cl.mat <- consensus_combine(cc, element = "class")
pvalue <- my_sigclust(x = dat, k = nk, nsim = 1000, labflag = 1, label =
cl.mat$`4`[, 1], icovest = 2)
str(pvalue)
}
