% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_cluster.R
\name{pl_cluster}
\alias{pl_cluster}
\title{Parallel consensus_cluster}
\usage{
pl_cluster(data, nk = 4, reps = 1000, algorithms = NULL,
  nmf.method = c("brunet", "lee"), distance = "euclidean",
  file.name = NULL, ...)
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}

\item{reps}{number of subsamples}

\item{algorithms}{vector of clustering algorithms for performing consensus
clustering. Must be any number of the following: "nmf", "hc", "diana",
"km", "pam", "ap", "sc", "gmm", "block", "som", "cmeans", "hdbscan". A
custom clustering algorithm can be used.}

\item{nmf.method}{specify NMF-based algorithms to run. By default the
"brunet" and "lee" algorithms are called. See \code{\link[NMF:nmf]{NMF::nmf()}} for details.}

\item{distance}{a vector of distance functions. Defaults to "euclidean".
Other options are given in \code{\link[stats:dist]{stats::dist()}}. A custom distance function can
be used.}

\item{file.name}{if not \code{NULL}, the returned array will be saved at each
iteration as well as at the end of the function call to an \code{rds} object
with \code{file.name} as the file name.}

\item{...}{other arguments passed to \code{consensus_cluster}}
}
\value{
one file for each combination of \code{algorithms}, \code{distance}, and
\code{reps}. For NMF, each \code{nmf.method} gets a separate file. Each object is a
4D array with one column, so that they can be run in parallel and combined
at the end.
}
\description{
Parallel consensus_cluster
}
\examples{
\dontrun{
data(hgsc)
pl_cluster(data, nk = 4, reps = 2, algorithms = c("nmf", "km", "block"),
distance = "manhattan", nmf.method = "lee", file.name = "test")
}
}
