% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_confun.R
\name{pl_confun}
\alias{pl_confun}
\title{Parallel consensus function}
\usage{
pl_confun(nk, cons.funs = c("kmodes", "majority", "CSPA", "LCE"),
  sim.mat = c("cts", "srs", "asrs"))
}
\arguments{
\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}

\item{cons.funs}{consensus functions to use. Current options are "kmodes"
(k-modes), "majority" (majority voting), "CSPA" (Cluster-based Similarity
Partitioning Algorithm), "LCE" (linkage clustering ensemble)}

\item{sim.mat}{similarity matrix; choices are "cts", "srs", "asrs".}
}
\value{
A vector of final consensus class assignments using one of the
consensus functions.
}
\description{
The consensus functions are passed from \code{dice}.
}
