% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_conmat.R
\name{pl_conmat}
\alias{pl_conmat}
\title{Parallel consensus matrix}
\usage{
pl_conmat()
}
\value{
one file for each "imputed" output object, this time prefixed with
"conmat".
}
\description{
Create a consensus matrix for each imputed cluster vector from the parallel
pipeline.
}
