% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_evaluate.R
\name{pl_evaluate}
\alias{pl_evaluate}
\title{Parallel evaluation of clustering results}
\usage{
pl_evaluate(data, nk, n = 5)
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}

\item{n}{an integer specifying the top \code{n} algorithms to keep after trimming
off the poor performing ones using Rank Aggregation. If the total number of
algorithms is less than \code{n} no trimming is done.}
}
\value{
Three objects: The internal evaluation indices matrix, a data frame
of cluster assignments for each of the algorithms and consensus functions,
and a data frame of the top \code{n} selected algorithms plus the \code{CC} and \code{Pr}
summaries.
}
\description{
Internal evaluation indices, cluster assignments, and summaries are returned.
}
\details{
The \code{CC} is the most commonly selected cluster. The \code{Pr} is the proportion at
which the cluster \code{CC} appears across different algorithms.
}
