% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_impute.R
\name{pl_impute}
\alias{pl_impute}
\title{Parallel K-Nearest Neighbours imputation}
\usage{
pl_impute(data, seed = 123)
}
\arguments{
\item{data}{data matrix}

\item{seed}{random seed for knn imputation reproducibility}
}
\value{
one file for each "raw" output object, this time prefixed with
"imputed".
}
\description{
Impute each individual clustering unit from \code{pl_cluster} and write results to
a new directory.
}
