% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_merge.R
\name{pl_merge}
\alias{pl_merge}
\title{Parallel cluster merging}
\usage{
pl_merge(data, nk = 4)
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}
}
\value{
A raw merged cluster array, a knn-imputed cluster array, and a
completely imputed (using majority voting) cluster array all saved to file
}
\description{
Produce the objects E, Eknn, and Ecomp from \code{dice} and save to file.
}
