% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_clusters.R
\name{relabel_clusters}
\alias{relabel_clusters}
\title{Relabel clusters}
\usage{
relabel_clusters(E)
}
\arguments{
\item{E}{N by M matrix of cluster ensemble}
}
\value{
A list with elements
\item{newE}{N by M matrix of relabelled ensemble}
\item{no_allcl}{total number of clusters in the ensemble}
}
\description{
Relabel clusters in the ensemble \code{E}
}
\examples{
data("E_LCE")
E_LCE_relabelled <- relabel_clusters(E_LCE)
}
\author{
Johnson Liu
}
\references{
MATLAB function relabelCl by Simon Garrett in LinkCluE package
}

