% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelCl.R
\name{relabelCl}
\alias{relabelCl}
\title{Title function used for relabelling clusters in the ensemble 'E'}
\usage{
relabelCl(E)
}
\arguments{
\item{E}{: N by M matrix of cluster ensemble}
}
\value{
newE: N by M matrix of relabelled ensemble, no_allcl: total number of clusters in the ensemble
}
\description{
Title function used for relabelling clusters in the ensemble 'E'
}
\examples{
data("E_LCE")
E_LCE_relabelled<-relabelCl(E_LCE)
}

