% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{similarityMatrices}
\alias{similarityMatrices}
\title{Similarity matrices for three clustering algorithms}
\format{A list of 3 matrices each of dimension 489 by 489.}
\usage{
similarityMatrices
}
\description{
Similarity matrices for three clustering algorithms
}
\keyword{datasets}

