% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortMatrixRowWise.R
\name{sortMatrixRowWise}
\alias{sortMatrixRowWise}
\title{Title function to rows of a matrix}
\usage{
sortMatrixRowWise(M, order)
}
\arguments{
\item{M}{: input matrix}

\item{order}{: "ascending" for ascending order or "descending" for descending order}
}
\value{
a matrix whose rows are sorted accordingly
}
\description{
Title function to rows of a matrix
}

