% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs}
\alias{srs}
\title{Title function for computing simrank based similarity matrix}
\usage{
srs(E, dc, R)
}
\arguments{
\item{E}{: matrix of cluster ensemble}

\item{dc}{: decay factor, [0,1]}

\item{R}{: number of iterations for simrank algorithm}
}
\description{
Title function for computing simrank based similarity matrix
}

