% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIJ.R
\name{UIJ}
\alias{UIJ}
\title{Title function to update U, I, and J for function linkage}
\usage{
UIJ(i, j, m)
}
\arguments{
\item{i}{: see i in linkage.R}

\item{j}{: see j in linkage.R}

\item{m}{: see m in linkage.R}
}
\value{
a list of U, I, and J in linkage.R
}
\description{
Title function to update U, I, and J for function linkage
}

