% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validilityScores.R
\name{valid_CA}
\alias{valid_CA}
\title{Title function for computing classification accuracy for a clustering result}
\usage{
valid_CA(labels, truelabels)
}
\arguments{
\item{labels}{: labels generated by classifier}

\item{truelabels}{: true labels for the observations}
}
\value{
classification accuracy score
}
\description{
Title function for computing classification accuracy for a clustering result
}
\examples{
valid_CA(c(4,4,4,4),c(4,2,4,4))
valid_CA(c(4,4,4,4,8),c(4,2,4,4,6))
}

