% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_RandIndex.R
\name{valid_RandIndex}
\alias{valid_RandIndex}
\title{Title function to calculate Rand indices to compare two partitions}
\usage{
valid_RandIndex(c1, c2)
}
\arguments{
\item{c1}{: vector listing class membership}

\item{c2}{: another vector listing class membership}
}
\value{
AR: adjusted Rand index, RI: unadjusted Rand index, MI: Mirkin's index, HI: Hubert's index
}
\description{
Title function to calculate Rand indices to compare two partitions
}

