% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validilityScores.R
\name{valid_sumsqures}
\alias{valid_sumsqures}
\title{Title function to compute within group, between group, and total sum of squares and cross-products}
\usage{
valid_sumsqures(data, labels, k)
}
\arguments{
\item{data}{: a matrix with each column representing a variable}

\item{labels}{: a vector indicating class labels}

\item{k}{: number of clusters}
}
\value{
W: within group sum of squares and cross-products; B: between group sum of squares and cross-products;T: total sum of squares and cross-products;
         Sintra & Sinter: centroid diameter and linkage distance
}
\description{
Title function to compute within group, between group, and total sum of squares and cross-products
}
\examples{
data("FGT")
data("FGD")
valid_sumsqures(data=FGD,labels=FGT,k=4)
}

