% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightCl.R
\name{weightCl}
\alias{weightCl}
\title{Title function to compute weight for each pair of clusters using their shared members (Jaccard coefficient)}
\usage{
weightCl(E)
}
\arguments{
\item{E}{: N by M matrix of cluster ensemble}
}
\value{
wcl: an p by p weighted cluster matrix where p denotes number of classes
}
\description{
Title function to compute weight for each pair of clusters using their shared members (Jaccard coefficient)
}
\examples{
data("E_LCE")
wl<-weightCl(E_LCE)
}

