% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asrs.R
\name{asrs}
\alias{asrs}
\title{Title A function for computing approximated simrank based similarity matrix}
\usage{
asrs(E, dc)
}
\arguments{
\item{E}{: N by M matrix of cluster ensemble}

\item{dc}{:decay factor, range 0 to 1, inclusive}
}
\value{
S: N by N ASRS matrix
}
\description{
Title A function for computing approximated simrank based similarity matrix
}
\examples{
data(E_LCE) 
ASRS<-asrs(E_LCE, 0.8)
}

