% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cluster.R
\name{clusternum}
\alias{clusternum}
\title{Title function to assign leaves under cluster c to c, equivalent to MATLAB function clusternum}
\usage{
clusternum(X, resultT, k, c)
}
\arguments{
\item{X:}{original hierarchical cluster tree}

\item{resultT:}{cluster tree generated in cluster}

\item{k:}{level in the tree}

\item{c:}{cluster to be assigned to the leaf}
}
\description{
Title function to assign leaves under cluster c to c, equivalent to MATLAB function clusternum
}

