% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMax.R
\name{colMax}
\alias{colMax}
\title{Function to find the maximum in each column of a matrix}
\usage{
colMax(M)
}
\arguments{
\item{M}{: a matrix}
}
\value{
vec: a vector containning the maximum in each column of input matrix
}
\description{
Function to find the maximum in each column of a matrix
}

