% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMin.R
\name{colMin}
\alias{colMin}
\title{Function to find the minimum in each column of a matrix}
\usage{
colMin(M)
}
\arguments{
\item{M}{: a matrix}
}
\value{
vec: a vector containning the minimum in each column of input matrix
}
\description{
Function to find the minimum in each column of a matrix
}
\examples{
colMin(matrix(c(1,2,3,4,5,6,7,8,9),ncol=3))
}

