% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conClustEnsemble.R
\name{ConClustEnsemble}
\alias{ConClustEnsemble}
\title{Generate a cluster ensemble using the K-means algorithm}
\usage{
ConClustEnsemble(X, k, reps = 1000)
}
\arguments{
\item{X}{a data set, rows are observations, columns are variables}

\item{k}{number of clusters in the base clusterings}

\item{M}{number of base clusterings in the ensemble}

\item{scheme}{cluster ensemble generating scheme (1=fixed k, 2=random k)}
}
\value{
matrix of cluster ensemble
}
\description{
Generate a cluster ensemble using the K-means algorithm
}
\author{
Johnson Liu
}

