% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConClustEnsemble.R
\name{ConClustEnsemble}
\alias{ConClustEnsemble}
\title{Function to perform clustering using \code{\link{ConClust}} with all default 
methods using user-sepcified number of clusters and number of repetitions.}
\usage{
ConClustEnsemble(X, k, reps = 1000, method = NULL)
}
\arguments{
\item{X}{data set with rows as samples, columns as genes}

\item{k}{number of clusters in the base clusterings}

\item{reps}{number of repetitions}

\item{method}{vector of clustering algorithms for performing consensus 
clustering. Must be any number of the following: "nmfDiv", "nmfEucl", 
"hcAEucl", "hcDianaEucl", "kmEucl", "kmSpear", "pamEucl", "pamSpear", 
"apEucl", "scRbf", "gmmBIC", "biclust".}
}
\value{
matrix of cluster ensemble
}
\description{
Function to perform clustering using \code{\link{ConClust}} with all default 
methods using user-sepcified number of clusters and number of repetitions.
}
\examples{
# Generate a 489 by 10 clustering ensemble using 2 algorithms in ConClust
data(hgsc)
dat <- t(hgsc[, -1])
colnames(dat) <- hgsc[, 1]
E <- ConClustEnsemble(X = dat, k = 4, reps = 5, method = c("hcAEucl", "kmEucl"))
}
\author{
Johnson Liu
}

