% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConClustEnsemble.R
\name{ConClustEnsemble}
\alias{ConClustEnsemble}
\title{Generate a cluster ensemble using the K-means algorithm}
\usage{
ConClustEnsemble(X, k, reps = 1000)
}
\arguments{
\item{X}{a data set, rows are observations, columns are variables}

\item{k}{number of clusters in the base clusterings}

\item{M}{number of base clusterings in the ensemble}
}
\value{
matrix of cluster ensemble
}
\description{
Function to perform clustering using ConClust::ConClust with default methods nmfDiv,nmfEucl,hcAEucl,hcDianaEucl, kmEucl,kmSpear,pamEucl,apEucl,
apEucl, scRbf, gmmBIC, and biclust with user-sepcified number of clusters and number of repetitions
}
\examples{
library(magrittr)
library(tidyr)
library(reshape2)
library(biostatUtil)
library(ConClust)
data("hgsc")
dat <- hgsc[,-1]
rownames(dat) <- hgsc[,1]
tdat <- t(dat)
mdat <- melt(tdat) \%>\%
 set_colnames(c("patientID","Gene","Expr"))
E<-ConClustEnsemble(X=tdat,k=4,reps=10) 
The example will generate a 489 by 120 clustering ensemble using the default 12 algorithms in ConClust
}
\author{
Johnson Liu
}

