% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_summary.R
\name{consensus_summary}
\alias{consensus_summary}
\title{Consensus summary}
\usage{
consensus_summary(E)
}
\arguments{
\item{E}{output from \code{consensus_cluster}}
}
\value{
A list with summaries for each algorithm. Each algorithm has a list 
  with two elements: consensus_matrix and consensus_class
}
\description{
Given an object from \code{\link{consensus_cluster}}, returns a list of
consensus matrices and consensus classes for each clustering algorithm.
}
\examples{
data(hgsc)
dat <- t(hgsc[, -1])
x <- consensus_cluster(dat, nk = 3:4, reps = 5, algorithms = "hc",
progress = FALSE)
cs <- consensus_summary(x)
str(cs)
}
\author{
Derek Chiu
}

