% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_summary.R
\name{consensus_summary}
\alias{consensus_summary}
\title{Consensus summary}
\usage{
consensus_summary(res, k = NULL, progress = TRUE, save = FALSE,
  file.name = "results_CC")
}
\arguments{
\item{res}{output from \code{ConClust}}

\item{progress}{logical; should a progress bar be displayed?}

\item{save}{logical; if \code{TRUE}, the returned object will be saved at
each iteration as well as at the end.}

\item{file.name}{file name of the written object}
}
\value{
A list with summaries for each algorithm. Each algorithm has a list
  with two elements: consensus_matrix and consensus_class
}
\description{
Given an object from \code{\link{ConClust}}, returns a list of consensus
matrices and consensus classes for each clustering algorithm.
}
\examples{
data(hgsc)
dat <- t(hgsc[, -1])
x <- ConClust(dat, nc = 2:4, reps = 10, method = c("hcAEucl", "kmEucl"))
y1 <- consensus_summary(x)
y2 <- consensus_summary(x, k = 4)
str(y1)
str(y2)
}
\author{
Derek Chiu
}
\seealso{
Other consensus functions: \code{\link{consensus_class}},
  \code{\link{consensus_combine}},
  \code{\link{consensus_confmat}},
  \code{\link{consensus_matrix}}
}

