% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_funs.R
\name{CSPA}
\alias{CSPA}
\title{Cluster-based Similarity Partitioning Algorithm (CSPA)}
\usage{
CSPA(x, k, method = "average")
}
\arguments{
\item{x}{a \code{\link{consensus_matrix}}}

\item{k}{number of clusters}

\item{method}{linkage type for hierarchical clustering. Defaults to
"average". See \code{\link[stats]{hclust}} for details.}
}
\value{
cluster assignments for the consensus class
}
\description{
Performs hierarchical clustering on a consensus matrix to obtain consensus
class labels.
}
\examples{
set.seed(2)
x <- replicate(100, rbinom(100, 4, 0.2))
cm <- consensus_matrix(x)
CSPA(cm, k = 3)
}
\author{
Derek Chiu
}
\seealso{
Other consensus functions: \code{\link{LCE}},
  \code{\link{k_modes}}, \code{\link{majority_voting}}
}

