% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_funs.R
\name{CSPA}
\alias{CSPA}
\title{Cluster-based Similarity Partitioning Algorithm (CSPA)}
\usage{
CSPA(E, k)
}
\arguments{
\item{E}{is an array of clustering results.}

\item{k}{number of clusters}
}
\value{
cluster assignments for the consensus class
}
\description{
Performs hierarchical clustering on a stack of consensus matrices to obtain
consensus class labels.
}
\examples{
data(hgsc)
dat <- t(hgsc[, -1])[1:100, 1:50]
x <- consensus_cluster(dat, nk = 4, reps = 4, algorithms = c("nmfEucl", 
"hcAEucl", "hcDianaEucl"), save = FALSE)
CSPA(x, k = 4)
}
\author{
Derek Chiu
}
\seealso{
Other consensus functions: \code{\link{LCE}},
  \code{\link{k_modes}}, \code{\link{majority_voting}}
}

