% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{A function that executes the diceR algorithm}
\usage{
dice(data, nk, R = 10, algorithms = c("hcAEucl", "kmEucl", "scRbf",
  "gmmBIC"), consensusFUNS = c("kmodes", "CSPA", "majority", "LCE"),
  prune = FALSE, reweigh = FALSE, evaluate = "internal",
  refClass = NULL)
}
\arguments{
\item{data}{a data set with rows as observations, columns as variables}

\item{nk}{a rangeof cluster sizes (or a single value).}

\item{R}{number of data subsamples to obtain. (See ConClust documentation)}

\item{algorithms}{clustering algorithms to be used in the ensemble. Current
options are c("nmfDiv", "nmfEucl", "hcAEucl", "hcDianaEucl",
"kmEucl","kmSpear", "pamEucl", "pamSpear", "apEucl", "scRbf",
"gmmBIC","biclust"). Please see ConClust for definition.}

\item{consensusFUNS}{Consensus Functions to use. Current options are majority
voting, k-modes, CSPA, LCE}

\item{prune}{a logical value to indicate whether algorithm pruning is needed
prior to consensus clustering. defaults to FALSE.}

\item{reweigh}{a logical value to indicate whether after pruning, certain
algorithms should have more weight than others. defaults to FALSE.If weigh
is TRUE pruning will be set to TRUE as well.}

\item{evaluate}{indicate whether internal evaluation is needed.
evaluate=c("internal","external","both"). If evaluate is either external or
both, then a reference class must be provided.}

\item{refClass}{a vector of length n, indicating the reference class to
compare against.}
}
\description{
(Needs to be more informative) A function that runs through the full diceR
algorithm
}

