% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Diverse Clustering Ensemble}
\usage{
dice(data, nk, reps = 10, algorithms = NULL, cons.funs = c("kmodes",
  "CSPA", "majority", "LCE"), sim.mat = c("cts", "srs", "asrs"),
  trim = FALSE, reweigh = FALSE, evaluate = TRUE, plot = FALSE,
  ref.cl = NULL, progress = TRUE)
}
\arguments{
\item{data}{matrix with rows as observations, columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}

\item{reps}{number of data subsamples to generate. See 
\code{\link{consensus_cluster}} for details.}

\item{algorithms}{clustering algorithms to be used in the ensemble. Current 
options are "nmfDiv", "nmfEucl", "hcAEucl", "hcDianaEucl", "kmEucl", 
"kmSpear", "pamEucl", "pamSpear", "apEucl", "scRbf", "gmmBIC", and 
"biclust". See \code{\link{consensus_cluster}} for details.}

\item{cons.funs}{consensus functions to use. Current options are "kmodes" 
(k-modes), "majority" (majority voting), "CSPA" (Cluster-based Similarity 
Partitioning Algorithm), "LCE" (linkage clustering ensemble)}

\item{sim.mat}{type of similarity matrix. One of "cts", "srs", "asrs. See 
\code{\link{LCE}} for details.}

\item{trim}{logical; if \code{TRUE}, the number of algorithms in 
\code{algorithms} is reduced based on internal validity index performance 
prior to consensus clustering by \code{cons.funs}. Defaults to 
\code{FALSE}.}

\item{reweigh}{logical; if \code{TRUE}, algorithms are reweighted based on 
internal validity index performance after trimming. Well-performing 
algorithms are given higher weight prior to consensus clustering by 
\code{cons.funs}. Defaults to \code{FALSE}. Ignored if \code{trim = FALSE}.}

\item{evaluate}{logical; if \code{TRUE} (default), validity indices are 
returned. Internal validity indices are always computed. If \code{ref.cl} 
is not \code{NULL}, then external validity indices will also be computed.}

\item{plot}{logical; if \code{TRUE}, \code{\link{graph_all}} is called and 
relevant graphs are outputted. Ignored if \code{evaluate = FALSE}.}

\item{ref.cl}{reference class; a vector of length equal to the number of 
observations.}

\item{progress}{logical; if \code{TRUE} (default), progress bar is shown.}
}
\value{
A final clustering assignment from the diverse clustering ensemble 
  method.
}
\description{
Runs consensus clustering across subsamples, algorithms, and number of 
clusters (k).
}
\examples{
library(dplyr)
data(hgsc)
dat <- t(hgsc[, -1])
ref.cl <- data.frame(initCol = rownames(dat)) \%>\%
tidyr::separate(initCol,
                into = c("patientID", "Class"),
                sep = "_") \%>\% 
  magrittr::use_series(Class) \%>\% 
  factor() \%>\% 
  as.integer()
dice.obj <- dice(dat, nk = 4, reps = 5, algorithms = c("hcAEucl", "pamEucl"),
cons.funs = c("kmodes", "majority"), ref.cl = ref.cl)
str(dice.obj)
}
\author{
Aline Talhouk, Derek Chiu
}

