% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.r
\name{dice}
\alias{dice}
\alias{\%d\%}
\title{Roll the dice}
\usage{
dice(r)

n \%d\% m
}
\arguments{
\item{r}{A character vector in the format "ndm" where \emph{n} is the number of
dice and \emph{m} is the number of faces.}

\item{n}{An integer representing the number of dice.}

\item{m}{An integer representing the number of faces per die.}
}
\value{
An integer vector with the class "dice".
}
\description{
These functions let us roll the dice
}
\details{
This is a description of how dice are made...
}
\examples{
dice("2d6")

2 \%d\% 6
}
\author{
Paul Egeler, MS
}
