% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dice.r
\name{fast_roll}
\alias{fast_roll}
\alias{dice}
\alias{\%d\%}
\title{Roll the dice}
\usage{
fast_roll(r)

dice(r)

n \%d\% m
}
\arguments{
\item{r}{A character vector in the format "ndm" where \emph{n} is the number of
dice and \emph{m} is the number of faces.}

\item{n}{An integer representing the number of dice.}

\item{m}{An integer representing the number of faces per die.}
}
\value{
An integer vector with the class "dice".
}
\description{
These functions let us roll the dice
}
\details{
The \code{fast_roll} function is an example
\href{https://cran.r-project.org/package=Rcpp}{Rcpp} function. See more in the
\emph{src/} folder of the package source.
}
\examples{
dice("2d6")

2 \%d\% 6
}
\author{
Paul Egeler, MS
}
