% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diceR-package.r
\docType{package}
\name{diceR-package}
\alias{diceR-package}
\title{diceR}
\description{
The diceR package performs Consensus Clustering, an ensemble clustering 
algorithm that pools together results from different clustering realizations.
Each realization uses a unique combination of observations and diverse
clustering algorithms.
}
