% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_imputed.R
\docType{data}
\name{E_imputed}
\alias{E_imputed}
\title{Example data for imputation}
\format{An array with 489 rows, 10 columns, 12 slices}
\usage{
E_imputed
}
\description{
The data set is intended to be tested with \code{\link{knn_impute}}
}
\keyword{datasets}

