% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_validity.R
\name{external_validity}
\alias{ev_accuracy}
\alias{ev_nmi}
\alias{ev_rand}
\alias{external_validity}
\title{External validity indices}
\usage{
ev_accuracy(pred.lab, ref.lab)

ev_rand(pred.lab, ref.lab)

ev_nmi(pred.lab, ref.lab, method = "emp")
}
\arguments{
\item{pred.lab}{predicted labels generated by classifier}

\item{ref.lab}{reference labels for the observations}

\item{method}{method of computing the entropy. Can be any one of "emp", "mm",
"shrink", or "sg".}
}
\value{
\code{ev_accuracy} returns the classification accuracy

\code{ev_rand} returns a list with elements
  \item{AR}{adjusted Rand index}
  \item{RI}{unadjusted Rand index}
  \item{MI}{Mirkin's index}
  \item{HI}{Hubert's index}
}
\description{
\strong{E}xternal \strong{v}alidity indices compare a predicted clustering 
result with a reference class or gold standard.
}
\details{
\code{ev_accuracy} computes the classification accuracy for a clustering
result

\code{ev_rand} calculates Rand indices (and their variants) to
  compare two partitions

\code{ev_nmi} calculates the normalized mutual information
}
\note{
\code{ev_nmi} is adapted from \code{\link[infotheo]{mutinformation}}
}
\examples{
set.seed(1)
E <- matrix(rep(sample(1:4, 1000, replace = TRUE)), nrow = 100, byrow =
              FALSE)
x <- sample(1:4, 100, replace = TRUE)
y <- sample(1:4, 100, replace = TRUE)
ev_accuracy(E[, 1], E[, 2])
ev_rand(E[, 1], E[, 2])
ev_nmi(x, y)
}
\author{
Johnson Liu, Derek Chiu
}
\references{
MATLAB function \code{valid_CA} and \code{valid_RandIndex} in
  package LinkCluE by Simon Garrett

Strehl A, Ghosh J. Cluster ensembles: a knowledge reuse framework
  for combining multiple partitions. J. Mach. Learn. Res. 2002;3:583-617.
}

