% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_validity.R
\name{external_validity}
\alias{ev_accuracy}
\alias{ev_rand}
\alias{external_validity}
\title{External validity indices}
\usage{
ev_accuracy(pred.lab, ref.lab)

ev_rand(pred.lab, ref.lab)
}
\arguments{
\item{pred.lab}{predicted labels generated by classifier}

\item{ref.lab}{reference labels for the observations}
}
\value{
\code{ev_accuracy} returns the classification accuracy

\code{ev_rand} returns a list with elements
  \item{AR}{adjusted Rand index}
  \item{RI}{unadjusted Rand index}
  \item{MI}{Mirkin's index}
  \item{HI}{Hubert's index}
}
\description{
\strong{E}xternal \strong{v}alidity indices compare a predicted clustering 
result with a reference class or gold standard.
}
\details{
\code{ev_accuracy} computes the classification accuracy for a clustering
result

\code{ev_rand} calculates Rand indices (and their variants)to
  compare two partitions
}
\examples{
data("E_LCE")
ev_accuracy(c(4, 4, 4, 4), c(4, 2, 4, 4))
ev_accuracy(c(4, 4, 4, 4, 8), c(4, 2, 4, 4, 6))
ev_rand(E_LCE[, 1], E_LCE[, 2])
}

