% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_cdf}
\alias{graph_cdf}
\title{Cumulative distribution function graphs}
\usage{
graph_cdf(x)
}
\arguments{
\item{x}{an object from \code{\link{ConClust}}}
}
\value{
a plot showing the cdfs for different algortihms
}
\description{
Graph CDF for consensus matrices from different algorithms
}
\examples{
# Consensus clustering for 3 algorithms
set.seed(911)
x <- matrix(rnorm(1000), nrow = 10)
CC1 <- ConClust(x, k = 4, reps = 10,
method = c("hcAEucl", "apEucl", "gmmBIC"), save = FALSE)

# Plot CDF
p <- graph_cdf(CC1)
p

# Change y label and add colours
p + labs(y = "Probability") + stat_ecdf(aes(colour = Method)) +
scale_color_brewer(palette = "Set2")

# Heatmaps with column side colours corresponding to clusters for each algorithm
graph_heatmap(CC1)
}
\author{
Derek Chiu
}

