% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graphs}
\alias{graph_cdf}
\alias{graph_delta_area}
\alias{graph_heatmap}
\alias{graph_tracking}
\alias{graphs}
\title{Graphical Displays}
\usage{
graph_cdf(x)

graph_delta_area(x)

graph_heatmap(x, main = NULL, ...)

graph_tracking(x)
}
\arguments{
\item{x}{an object from \code{\link{ConClust}}}

\item{main}{heatmap title}

\item{...}{additional arguments to \code{\link[gplots]{heatmap.2}}}
}
\value{
Various plots from \code{graph_*{}} functions. All plots are 
  generated using \code{ggplot}, except for \code{graph_heatmap}, which uses 
  \code{\link[gplots]{heatmap.2}}. Colours used in \code{graph_heatmap} and
  \code{graph_tracking} utilize \code{\link{RColorBrewer}} palettes.
}
\description{
Graph cumulative distribution function (CDF) graphs, relative change in area 
under CDF curves, heatmaps, and cluster assignment tracking plots.
}
\details{
\code{graph_cdf} plots the CDF for consensus matrices from different 
algorithms. \code{graph_delta_area} calculates the relative change in area 
under CDF curve between algorithms. \code{graph_heatmap} generates consensus
matrix heatmaps for each algorithm in \code{x}. \code{graph_tracking} tracks
how cluster assignments change between algorithms.
}
\note{
The current implementations of \code{graph_delta_area} and
  \code{graph_tracking} do not compare between different number of clusters.
  Thus the interpretation of relative change is not natural, and the tracking
  plot shows low concordance.
}
\examples{
# Consensus clustering for 3 algorithms
set.seed(911)
x <- matrix(rnorm(1000), nrow = 10)
CC1 <- ConClust(x, k = 4, reps = 10,
method = c("hcAEucl", "apEucl", "gmmBIC"), save = FALSE)

# Plot CDF
p <- graph_cdf(CC1)
p

# Change y label and add colours
p + labs(y = "Probability") + stat_ecdf(aes(colour = Method)) +
scale_color_brewer(palette = "Set2")

# Delta Area
p <- graph_delta_area(CC1)
p

# Heatmaps with column side colours corresponding to clusters for each algorithm
graph_heatmap(CC1)

# Track how cluster assignments change between algorithms
p <- graph_tracking(CC1)
p
}
\author{
Derek Chiu
}

