% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgsc}
\alias{hgsc}
\title{High grade serous cancer data from TCGA used to classify subtypes}
\format{A data frame with 321 rows and 490 variables}
\usage{
hgsc

hgsc
}
\description{
There are 321 genes and 489 samples. The first column is a gene identifier.

The 1st dimension indicates number of samples (489), second dimension is number of repetitions (50), third dimension is number of algorithms
}
\keyword{datasets}

