% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_knn}
\alias{impute_knn}
\title{K-Nearest Neighbours imputation
Set NA cases as test, otherwise training; neighbours = 5, min vote = 3}
\usage{
impute_knn(x, data, seed)
}
\arguments{
\item{x}{clustering object}

\item{data}{data matrix}

\item{seed}{random seed for knn imputation}
}
\description{
K-Nearest Neighbours imputation
Set NA cases as test, otherwise training; neighbours = 5, min vote = 3
}
