% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values}
\usage{
impute_missing(E4, data, imputeALL = TRUE, seed = 123456)
}
\arguments{
\item{E4}{is an array of clusterings with number of rows equal to the number 
of cases to be clustered, number of columns equal to the clustering 
obtained by different resampling of the data, and the third dimension is 
the different algorithms and the fourth dimension is the cluster size.}

\item{data}{is data matrix with samples as rows and genes/features as columns}

\item{imputeALL}{if \code{FALSE} the function will only call \code{impute_knn} to 
impute NAs missing due to resampling; if \code{TRUE}, the function will use
\code{impute_knn} and majority voting to eliminate NAs and relabel classes}

\item{seed}{random seed for KNN reproducibility}
}
\value{
E_imputed a list with a list an array of similar dimension to E with 
  elements that were left out imputed using KNN, as well as a flat matrix of 
  clusterings (E_imputed2), fully imputed and relabelled, if imputeALL was 
  set to TRUE.
}
\description{
Impute missing values from bootstrapped subsampling
}
\details{
The default output from \code{ConClust} will undoubtedly contain \code{NA} 
entries because each replicate chooses a random subset (with replacement) of 
all samples. Missing values are first imputed using KNN (K-Nearest 
Neighbours), with the non-missing cases indicating the training set, and
missing cases indicating the test set. Not all missing values are guaranteed
to be imputed by KNN. See \code{\link[class]{knn}} for details. Thus, any
remaining missing values are imputed using majority voting.
}
\note{
We consider 5 nearest neighbours and the minimum vote for definite 
  decision is 3.
}
\examples{
data(hgsc)
data <- t(hgsc[, -1])[1:100, 1:50]
E <- ConClust(data, nc = 3:4, reps = 10, method = c("hcAEucl", "kmEucl",
"scRbf"))
sum(is.na(E))
sum(is.na(impute_missing(E, data, imputeALL = TRUE)$E_imputed))
sum(is.na(impute_missing(E, data, imputeALL = TRUE)$E_imputed2))
}
\author{
Aline Talhouk
}

