% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values due to subsampling}
\usage{
imputeMissing(E, data, imputeALL = TRUE)
}
\arguments{
\item{E}{a matrix of clusterings with number of rows equal to the number of
cases to be clustered, number of columns equal to the clustering obtained
by different resampling of the data, and the third dimension is the
different algorithms.}

\item{data}{is data matrix with samples as rows and genes/features as columns}

\item{imputeALL}{if \code{FALSE} the function will only call knn_impute to
impute some NAs; if \code{TRUE}, the function will use knn_impute and
majority voting to eliminate NAs and relabel class}
}
\value{
a flat matrix of clusterings, fully imputed and relabelled.
}
\description{
Uses k nearest neighbour to impute missing values due to sampling. Not all
cases will be imputed using this method.
}
\examples{
data(hgsc)
data <- t(hgsc[,-1])[1:200, 1:100]
E <- ConClust(data, nc = 2:4, reps = 10, method = c("hcAEucl", "kmEucl",
"scRbf"))
imputeMissing(E, data)
}
\author{
Aline Talhouk
}

