% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissing.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values due to subsampling}
\usage{
imputeMissing(E, data, imputeALL = TRUE)
}
\arguments{
\item{data}{is data matrix with samples as rows and genes/features as columns}
}
\value{
a flat matrix of clusterings, fully imputed and relabelled.
}
\description{
Uses k nearest neighbour to impute missing values due to sampling, not all cases will be imputed using this method. If imputeALL is set to true (default), then cases will be relabelled and missing cases will be called by majority vote
}
\details{
#' @param E a matrix of clusterings with number of rows equal to the number of cases to be clustered, number of columns equal to the clustering obtained by different resampling of the data, and the third dimension is the different algorithms.
}
\author{
Aline Talhouk
}

