% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validity.R
\name{internal_validity}
\alias{internal_validity}
\alias{iv_compactness}
\alias{iv_db_dunn}
\alias{iv_sumsq}
\title{Internal validity indices}
\usage{
iv_compactness(data, labels)

iv_db_dunn(data, labels)

iv_sumsq(data, labels, k)
}
\arguments{
\item{data}{a dataset with rows as observations, columns as variables}

\item{labels}{a vector of cluster labels from a clustering result}

\item{k}{number of clusters}
}
\value{
\code{iv_compactness} returns the compactness score

\code{iv_db_dunn} returns a list with elements
  \item{DB}{Davies-Bouldin score}
  \item{Dunn}{Dunn score}

\code{iv_sumsq} returns a list with elements 
  \item{W}{within group sum of squares and cross-products}
  \item{B}{between group sum of squares and cross-products}
  \item{T}{total sum of squares and cross-products}
  \item{Sintra}{centroid diameter}
  \item{Sinter}{linkage distance}
}
\description{
\strong{I}nternal \strong{v}alidity indices are agnostic to any reference 
clustering results. They calculate cluster performance on the basis of 
compactness and separability.
}
\details{
\code{iv_compactness} computes the compactness validity index for a 
clustering result

\code{iv_db_dunn} computes the Davies-Bouldin index and Dunn index

\code{iv_sumsq} computes the within group, between group, and total
  sum of squares and cross-products
}
\examples{
data("FGD", "E_LCE", "FGT", "FGD")
iv_compactness(FGD, E_LCE[, 1])
iv_db_dunn(FGD, FGT)
iv_sumsq(data = FGD, labels = FGT, k = 4)
}
\references{
MATLAB functions \code{valid_compactness} and \code{valid_DbDunn}
  and \code{valid_sumsqures} by Simon Garrett in LinkCluE
}

