% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validity.R
\name{internal_validity}
\alias{internal_validity}
\alias{iv_chi}
\alias{iv_compactness}
\alias{iv_db_dunn}
\title{Internal validity indices}
\usage{
iv_compactness(data, labels)

iv_db_dunn(data, labels)

iv_chi(data, ...)
}
\arguments{
\item{data}{a dataset with rows as observations, columns as variables}

\item{labels}{a vector of cluster labels from a clustering result}

\item{...}{additional arguments to \code{\link[clusterSim]{index.G1}}}
}
\value{
\code{iv_compactness} returns the compactness score

\code{iv_db_dunn} returns a list with elements
  \item{DB}{Davies-Bouldin score}
  \item{Dunn}{Dunn score}

\code{iv_chi} returns the CHI
}
\description{
\strong{I}nternal \strong{v}alidity indices are agnostic to any reference 
clustering results. They calculate cluster performance on the basis of 
compactness and separability.
}
\details{
\code{iv_compactness} computes the compactness validity index for a 
clustering result

\code{iv_db_dunn} computes the Davies-Bouldin index and Dunn index

\code{iv_chi} computes the Calinski-Harabasz Index, or pseudo
  F-statistic. The function is a wrapper for
  \code{\link[clusterSim]{index.G1}}
}
\examples{
set.seed(1)
E <- matrix(rep(sample(1:4, 1000, replace = TRUE)), nrow = 100, byrow =
              FALSE)
set.seed(1)
dat <- as.data.frame(matrix(runif(1000, -10, 10), nrow = 100, byrow = FALSE))
iv_compactness(dat, E[, 1])
iv_db_dunn(dat, E[, 1])
iv_chi(dat, E[, 1], centrotypes = "centroids")
}
\references{
MATLAB functions \code{valid_compactness}, \code{valid_DbDunn}, 
  \code{valid_sumsqures} by Simon Garrett in LinkCluE

Calinski, T., Harabasz, J. (1974), A dendrite method for cluster
  analysis, "Communications in Statistics", vol. 3, 1-27.
}

