% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCE.R
\name{LCE}
\alias{LCE}
\title{Linkage Clustering Ensemble}
\usage{
LCE(E, data, k, dcCTS = 0.8, dcSRS = 0.8, dcASRS = 0.8, R = 10,
  sim.mat = c("cts", "srs", "asrs"))
}
\arguments{
\item{E}{is an array of clustering results. Missing entries are imputed using
\code{imputeMissing}}

\item{data}{original data matrix with rows as samples, columns as variables}

\item{k}{requested number of clusters}

\item{dcCTS}{decay constant for CTS matrix}

\item{dcSRS}{decay constant for SRS matrix}

\item{dcASRS}{decay constant for ASRS matrix}

\item{R}{number of repetitions for SRS matrix}

\item{sim.mat}{vector of similarity matrices to compute. Choices are "cts",
"srs", "asrs".}
}
\value{
a vector containing the cluster assignment from either the CTS, SRS,
  or ASRS similarity matrices
}
\description{
Generate CTS, SRS, and ASRS similarity matrices.
}
\examples{
data(hgsc)
dat <- t(hgsc[, -1])[1:100, 1:50]
x <- ConClust(dat, nc = 4, reps = 4,
              method = c("nmfEucl", "hcAEucl", "hcDianaEucl"), save = FALSE)
LCE(E = x, data = dat, k = 4, dcCTS = 0.8, dcSRS = 0.8, dcASRS = 0.8, R = 10,
sim.mat = "asrs")
}
\author{
Johnson Liu
}

