% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCE.R
\name{LCE}
\alias{LCE}
\title{Generate CTS, SRS, and ASRS similarity matrices.}
\usage{
LCE(E, dcCTS = 0.8, dcSRS = 0.8, dcASRS = 0.8, R = 10,
  is.relabelled = TRUE)
}
\arguments{
\item{E}{matrix of clustering results. Complete and relabelled cases needed.}

\item{dcCTS}{decay constant for CTS matrix}

\item{dcSRS}{decay constant for SRS matrix}

\item{dcASRS}{decay constant for ASRS matrix}

\item{R}{number of repetitions for SRS matrix}
}
\value{
a list containing the CTS, SRS, and ASRS matrix
}
\description{
Generate CTS, SRS, and ASRS similarity matrices.
}
\examples{
set.seed(1)
E <- matrix(rep(sample(1:4, 1000, replace = TRUE)), nrow = 100, byrow = FALSE)
link_clust(E = E, dcCTS = 0.8, dcSRS = 0.8, dcASRS = 0.8, R = 10,
           is.relabelled = FALSE)
data(hgsc)
dat <- t(hgsc[, -1])
x <- ConClust(dat[1:100, 1:50], k = 4, reps = 4,
              method = c("nmfEucl", "hcAEucl", "hcDianaEucl"), save = FALSE)
y <- link_clust(E = x, dcCTS = 0.8, dcSRS = 0.8, dcASRS = 0.8, R = 10)
}
\author{
Johnson Liu
}

