% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage}
\alias{linkage}
\title{Hierarchical clustering tree with different linkage types}
\usage{
linkage(S, method)
}
\arguments{
\item{S}{N by N similarity matrix}

\item{method}{linkage method; can be "average", "single", or "complete"}
}
\value{
matrix denoting hierarchical cluster tree
}
\description{
Create a hierarchical clustering tree using average, single, or complete
linkage.
}
\examples{
set.seed(1)
E <- matrix(rep(sample(1:4, 1000, replace = TRUE)), nrow = 100, byrow =
              FALSE)
s <- cts(E, 0.8)
lk <- linkage(s, "average")
}
\author{
Johnson Liu
}
\references{
function linkage in MATLAB for average, single, and complete
  linkage
}

