% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_voting.R
\name{majority_voting}
\alias{majority_voting}
\title{Combine clustering results using majority voting}
\usage{
majority_voting(E, is.relabelled = TRUE)
}
\arguments{
\item{E}{a matrix of clusterings with number of rows equal to the number of cases to be clustered, number of columns equal to the clustering obtained by different resampling of the data, and the third dimension is the different algorithms.}

\item{is.relabelled}{TRUE/FALSE defaults to TRUE, but if FALSE the data will be relabelled}
}
\value{
a vector of cluster assignment based on majority vote
}
\description{
Combine clustering results generated using different algorithms and different data perturbations
}
\examples{
data(E_imputed)
table(majority_voting(E_imputed, is.relabelled=FALSE))
}
\author{
Aline Talhouk
}

