% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_merge.R
\name{pl_merge}
\alias{pl_merge}
\alias{pl_merge_clust}
\alias{pl_merge_conmat}
\title{Parallel cluster and consensus matrix merging}
\usage{
pl_merge_clust(data, nk = 4)

pl_merge_conmat()
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{nk}{number of clusters (k) requested; can specify a single integer or a
range of integers to compute multiple k}
}
\value{
\code{pl_merge_clust} returns a raw merged cluster array, a knn-imputed
cluster array, and a completely imputed (using majority voting) cluster
array all saved to file. \code{pl_merge_conmat} returns a list of consensus
matrices.
}
\description{
\code{pl_merge_clust} constructs the objects E, Eknn, and Ecomp from \code{dice} and
saves to file. \code{pl_merge_conmat} constructs consensus matrices for each
algorithm and saves to file.
}
