% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cluster.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for consensus clustering}
\usage{
prepare_data(data, min.sd = 1)
}
\arguments{
\item{data}{data matrix with rows as samples and columns as variables}

\item{min.sd}{minimum standard deviation threshold. See details.}
}
\value{
dataset prepared for usage in \code{consensus_cluster}
}
\description{
Remove variables with low signal and scale before consensus clustering
}
\details{
The \code{min.sd} argument is used to filter the feature space for only
highly variable features. Only features with a standard deviation across all
samples greater than \code{min.sd} will be used.
}
\examples{
set.seed(2)
x <- replicate(10, rnorm(100))
prepare_data(x)
}
\author{
Derek Chiu
}

