% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{asrs}
\alias{cts}
\alias{similarity}
\alias{srs}
\title{Similarity Matrices}
\usage{
srs(E, dc, R)

asrs(E, dc)

cts(E, dc)
}
\arguments{
\item{E}{an N by M matrix of cluster ensembles}

\item{dc}{decay factor, ranges from 0 to 1 inclusive}

\item{R}{number of iterations for simrank algorithm}
}
\value{
an N by N SRS, ASRS, or CTS matrix
}
\description{
\code{srs} computes the simrank based similarity matrix, \code{asrs} computes
the approximated simrank based similarity matrix, and \code{cts} computes the
connected triple based similarity matrix
}
\examples{
data(E_LCE)
SRS <- srs(E_LCE, 0.8, 3)
ASRS <- asrs(E_LCE, 0.8)
CTS <- cts(E_LCE, 0.8)
}
\author{
Johnson Liu
}
\references{
MATLAB function srs, cts, asrs in package LinkCluE by Simon Garrett
}

