% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs}
\alias{srs}
\title{Title function for computing simrank based similarity matrix}
\usage{
srs(E, dc, R)
}
\arguments{
\item{E}{: N by M matrix of cluster ensemble}

\item{dc}{: decay factor, [0,1]}

\item{R}{: number of iterations for simrank algorithm}
}
\value{
N by N SRS matrix
}
\description{
Title function for computing simrank based similarity matrix
}
\examples{
data("E_LCE")
SRS<-srs(LCE,0.8,5)
}

