% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stod.R
\name{stod}
\alias{stod}
\title{Title converts similarity values to distance values and change matrix's format from square to vector (input format for linkage function)}
\usage{
stod(S)
}
\arguments{
\item{S}{: N by N similarity matrix}
}
\value{
d: distance vector
}
\description{
Title converts similarity values to distance values and change matrix's format from square to vector (input format for linkage function)
}

