% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stod.R
\name{stod}
\alias{stod}
\title{Similarity to Distance}
\usage{
stod(S)
}
\arguments{
\item{S}{N by N similarity matrix}
}
\value{
distance vector
}
\description{
Converts similarity values to distance values and change matrixs format from
square to vector (input format for linkage function)
}
\examples{
set.seed(1)
E<-matrix(rep(sample(1:4,1000,replace = TRUE)),nrow=100,byrow=FALSE)
d <- stod(cts(E,0.8))
}
\author{
Johnson Liu
}
\references{
MATLAB function stod in package LinkCluE by Simon Garrett
}

