% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stod.R
\name{stod}
\alias{stod}
\title{Similarity to Distance}
\usage{
stod(S)
}
\arguments{
\item{S}{N by N similarity matrix}
}
\value{
distance vector
}
\description{
Converts similarity values to distance values and change matrixs format from
square to vector (input format for linkage function)
}
\examples{
data("E_LCE")
ASRS <- asrs(E_LCE,0.8)
d <- stod(ASRS)
}
\author{
Johnson Liu
}

