% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_CA.R
\name{valid_CA}
\alias{valid_CA}
\title{Title function for computing classification accuracy for a clustering result}
\usage{
valid_CA(labels, truelabels)
}
\arguments{
\item{labels}{}

\item{truelabels}{}
}
\value{
classification accuracy score
}
\description{
Title function for computing classification accuracy for a clustering result
}

