% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validilityScores.R
\name{valid_compactness}
\alias{valid_compactness}
\title{Title function for computing compactness validity index for a clustering result}
\usage{
valid_compactness(X, labels)
}
\arguments{
\item{X}{: a dataset, rows are observations, columns are variables}

\item{labels}{: cluster labels from a clustering result(a vector)}
}
\value{
compactness score
}
\description{
Title function for computing compactness validity index for a clustering result
}
\examples{
data("FGD")
data("E")
labels<-E[,1]
valid_compactness(FGD,labels)
}

