% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validity.R
\name{valid_DbDunn}
\alias{valid_DbDunn}
\title{Title function for computing Davies-Bouldin index and Dunn index}
\usage{
valid_DbDunn(X, labels)
}
\arguments{
\item{X}{: a data set whose rows are observations, columns are variables}

\item{labels}{: cluster labels from a clustering result(a vector)}
}
\value{
DB: Davies-Bouldin score, Dunn: Dunn score
}
\description{
Title function for computing Davies-Bouldin index and Dunn index
}
\examples{
data("FGT")
data("FGD")
valid_DbDunn(FGD,FGT)
}

