% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigh_clusters.R
\name{weigh_clusters}
\alias{weigh_clusters}
\title{Weigh clusters}
\usage{
weigh_clusters(E)
}
\arguments{
\item{E}{N by M matrix of cluster ensembles}
}
\value{
a p by p weighted cluster matrix where p denotes number of classes
}
\description{
Compute weight for each pair of clusters using their shared members (Jaccard
coefficient)
}
\examples{
data("E_LCE")
wl <- weigh_clusters(E_LCE)
}
\author{
Johnson Liu
}
\references{
MATALAB function weightCl by Simon Garrett in package LinkCluE
}

