% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightCl.R
\name{weightCl}
\alias{weightCl}
\title{Title function to compute weight for each pair of clusters using their shared members (Jaccard coefficient)}
\usage{
weightCl(E)
}
\arguments{
\item{E}{: N by M matrix of cluster ensemble}
}
\value{
wcl: an weighted cluster matrix
}
\description{
Title function to compute weight for each pair of clusters using their shared members (Jaccard coefficient)
}

